% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_temp.R
\name{base_temp}
\alias{base_temp}
\title{Title}
\usage{
base_temp(tavg, d, type)
}
\arguments{
\item{tavg}{}

\item{d}{}

\item{type}{}
}
\value{

}
\description{
Title
}
\references{
Yang, S., Logan, J., & Coffey, D. L. (1995). Mathematical formulae for calculating the base temperature for growing degree days. In Agricultural and Forest Meteorology (Vol. 74, Issues 1-2, pp. 61-74). Elsevier BV
}
