% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollen_count.R
\docType{data}
\name{pollen_count}
\alias{pollen_count}
\title{Pollen count of alder, birch, and hazel}
\format{A data frame with 8352 rows and 5 variables:
\itemize{
    \item site
    \item date
    \item alder
    \item birch
    \item hazel
}}
\description{
pollen_count A dataset containing a synthetic data of alder, birch, and hazel pollen count in four locations ('Oz', 'Shire', 'Atlantis', 'Hundred Acre Wood') between 2007 and 2016
}
