% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollen-package.R
\docType{package}
\name{pollen-package}
\alias{pollen}
\alias{pollen-package}
\title{pollen: Analysis of Aerobiological Data}
\description{
Supports analysis of aerobiological data. Available features include determination of pollen season limits, replacement of outliers (Kasprzyk and Walanus (2014) <doi:10.1007/s10453-014-9332-8>), calculation of growing degree days (Baskerville and Emin (1969) <doi:10.2307/1933912>), and determination of the base temperature for growing degree days (Yang et al. (1995) <doi:10.1016/0168-1923(94)02185-M).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nowosad.github.io/pollen/}
  \item Report bugs at \url{https://github.com/Nowosad/pollen/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
