% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollen_season.R
\name{pollen_season}
\alias{pollen_season}
\title{A Pollen Season Function}
\usage{
pollen_season(x, value, date, method)
}
\arguments{
\item{x}{A data.frame with dates and pollen count values}

\item{value}{The name of the column with pollen count values}

\item{date}{The name of the dates column}

\item{method}{The pollen season method - "90", "95", "98", "Mesa", "Jager", or "Lejoly"}
}
\value{
A data.frame object with year, date of pollen season start and date of pollen season end
}
\description{
This function calculates the start and the end of pollen season for each year
}
\examples{

data(pollen_count)
df <- subset(pollen_count, site=='Oz')
pollen_season(df, value="birch", date="date", method="95")

df2 <- subset(pollen_count, site=='Atlantis')
pollen_season(df2, value="alder", date="date", method="95")

library('purrr')
pollen_count \%>\% split(., .$site) \%>\% 
                map(~pollen_season(., value="hazel", date="date", method="95"))
                
}
\references{
Nilsson S. and Persson S.: 1981, Tree pollen spectra in the Stockholm region (Sweden) 1973-1980, Grana 20, 179-182.

Andersen T.B.: 1991, A model to predict the beginning of the pollen season, Grana 30, 269-275.

Torben B.A.: 1991, A model to predict the beginning of the pollen season, Grana 30, 269-275.

Galan C., Emberlin J., Dominguez E., Bryant R.H. and Villamandos F.: 1995, A comparative analysis of daily variations in the Gramineae pollen counts at Cordoba, Spain and London, UK, Grana 34, 189-198.

Sanchez-Mesa J.A., Smith M., Emberlin J., Allitt U., Caulton E. and Galan C.: 2003, Characteristics of grass pollen seasons in areas of southern Spain and the United Kingdom, Aerobiologia 19, 243-250.

Jager S., Nilsson S., Berggren B., Pessi A.M., Helander M. and Ramfjord H.: 1996, Trends of some airborne tree pollen in the Nordic countries and Austria, 1980-1993. A comparison between Stockholm, Trondheim, Turku and Vienna, Grana 35, 171-178.

Lejoly-Gabriel and Leuschner: 1983, Comparison of air-borne pollen at Louvain-la-Neuve (Belgium) and Basel (Switzerland) during 1979 and 1980, Grana 22, 59-64.
}
\keyword{pollen}
\keyword{pollen,}
\keyword{season}

